rds: remove optional rds_subnet parameter
This parameter was used for the db_instance, but another group was
created if this parameter wasn't specified, however the rds_subnet
wouldn't be used by the db_instance itself.

This variable being removed, the db_subnet is always created, and always
used by the database, based on the private subnet provided.